better init check if file already exists

+++ bin/init/index.ts
@@ -1,7 +1,7 @@
 import packageJson from "../../package.json";
 import { join } from "path";
 import Paths from "../../src/paths";
-import { cpSync } from "fs";
+import { cpSync, existsSync } from "fs";
 import { webToken } from "@shpaw415/webtoken";
 
 export const PATH_TO_FRAME_MASTER = join(
@@ -28,24 +28,39 @@ async function init() {
   console.log("frame-master has been initialized in your project.");
 }
 
-function copyConfigFileToProject() {
+async function copyConfigFileToProject() {
   const defaultConfigFile = Bun.file(PATH_TO_DEFAULT_CONFIG_FILE);
   const targetPath = join(process.cwd(), Paths.configFile);
-  return Bun.file(targetPath).write(defaultConfigFile);
+  const targetFile = Bun.file(targetPath);
+  if (await targetFile.exists()) {
+    console.warn(`${Paths.configFile} already exists. Skipping copy.`);
+    return;
+  }
+  return targetFile.write(defaultConfigFile);
 }
 
 function copyDotFrameMasterDirToProject() {
   const targetDirPath = join(process.cwd(), ".frame-master");
+  if (existsSync(targetDirPath)) {
+    console.warn(`.frame-master directory already exists. Skipping copy.`);
+    return;
+  }
   cpSync(join(PATH_TO_FRAME_MASTER, ".frame-master"), targetDirPath, {
     recursive: true,
   });
 }
 
-function copyBunfigToProject() {
+async function copyBunfigToProject() {
   const bunfigFile = Bun.file(
     join(PATH_TO_FRAME_MASTER, "bin", "init", "bunfig.toml")
   );
+
   const targetPath = join(process.cwd(), "bunfig.toml");
+  const targetFile = Bun.file(targetPath);
+  if (await targetFile.exists()) {
+    console.warn(`bunfig.toml already exists. Skipping copy.`);
+    return;
+  }
   return Bun.file(targetPath).write(bunfigFile);
 }
 
@@ -56,9 +71,19 @@ async function addScriptsToPackageJson() {
   const packageJson = JSON.parse(packageJsonText);
 
   packageJson.scripts ??= {};
-  packageJson.scripts["dev"] =
-    "NODE_ENV=development bun --hot frame-master dev";
-  packageJson.scripts["start"] = "NODE_ENV=production bun frame-master start";
+  if (packageJson.scripts["dev"]) {
+    console.warn(
+      `"dev" script already exists in package.json. Skipping addition.`
+    );
+  } else
+    packageJson.scripts["dev"] =
+      "NODE_ENV=development bun --hot frame-master dev";
+  if (packageJson.scripts["start"]) {
+    console.warn(
+      `"start" script already exists in package.json. Skipping addition.`
+    );
+  } else
+    packageJson.scripts["start"] = "NODE_ENV=production bun frame-master start";
 
   return packageJsonFile.write(JSON.stringify(packageJson, null, 2));
 }
@@ -68,7 +93,7 @@ async function setEnvFile() {
   const envFileText = (await envFile.exists()) ? await envFile.text() : "";
 
   const modifiedText = [
-    envFileText,
+    ...envFileText.split("\n"),
     !envFileText.includes("WEB_TOKEN_IV=") &&
       `WEB_TOKEN_IV=${webToken.generateSecureIV()}`,
     !envFileText.includes("WEB_TOKEN_SECRET=") &&
